Use alternative mechanism for diffing cljs
This patch introduces an alternative mechanism for ClojureScript diffing. Given
the differences in how vars and macros work in Cljs, the current way is brittle
and also not compatible with self-host solutions like lumo.

The downside of this approach is that the bonus feature - throwing in case of
one argument only to clojure.test/is - is dropped.